+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The standard for the caret can [be found here](https://docs.npmjs.com/misc/semve
 Breaking changes result in a different major. UI changes that might break customizations on top of the sdk, will be treated as breaking changes too.
 
 
+## [new-version]
+
+### Added
+- Public: An error message is now shown if the upload file has as unsupported file type.
+
+
 ## [0.8.1]
 
 ## Fixed
@@ -72,6 +78,7 @@ Breaking changes result in a different major. UI changes that might break custom
 
 The standard for this change log can be found [here](http://keepachangelog.com/).
 
+[new-version]: https://github.com/onfido/onfido-sdk-ui/compare/0.8.0...master
 [0.8.1]: https://github.com/onfido/onfido-sdk-ui/compare/0.8.0...0.8.1
 [0.8.0]: https://github.com/onfido/onfido-sdk-ui/compare/0.7.0...0.8.0
 [0.7.0]: https://github.com/onfido/onfido-sdk-ui/compare/0.6.1...0.7.0